Use lombok @Delegate for delegating calls to dropwizard Snapshot
* Use lombok @Delegate for delegating calls to dropwizard Snapshot in
DropwizardMetricsStatistics

+++ src/main/java/org/obd/metrics/statistics/MetricStatistics.java
@@ -3,7 +3,7 @@ package org.obd.metrics.statistics;
 public interface MetricStatistics {
 	int size();
 
-	long getMedian();
+	double getMedian();
 
 	long getMax();
